Add virtual machine offer show method
The `status` method was replaced with `show` because the configured
endpoint it is hitting in the method is getting a specific offer, not
using the dedicated status endpoint.
The CLI arguments have been updated to allow for the config json
file to be passed in by the user. The VM offer functions are using
config file to get information to send to the API, so it should be an
option to be able to provide it.
Other comments have been updated to reflect the correct offer type.
This change also supports the unit tests in passing.